Extract the cli based log store as a generic impl for all OS
