+++ core/invoker/src/main/scala/org/apache/openwhisk/core/containerpool/docker/DockerForMacContainerFactory.scala
@@ -17,20 +17,11 @@
 
 package org.apache.openwhisk.core.containerpool.docker
 
-import java.time.Instant
-
 import akka.actor.ActorSystem
-import org.apache.openwhisk.common.{AkkaLogging, Logging, TransactionId}
+import org.apache.openwhisk.common.{Logging, TransactionId}
 import org.apache.openwhisk.core.WhiskConfig
 import org.apache.openwhisk.core.containerpool._
-import org.apache.openwhisk.core.containerpool.logging.{DockerToActivationLogStore, LogStore, LogStoreProvider}
-import org.apache.openwhisk.core.entity.{
-  ActivationLogs,
-  ExecutableWhiskAction,
-  Identity,
-  InvokerInstanceId,
-  WhiskActivation
-}
+import org.apache.openwhisk.core.entity.InvokerInstanceId
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration._
@@ -61,8 +52,6 @@ class DockerForMacClient(dockerHost: Option[String] = None)(executionContext: Ex
     with DockerApiWithFileAccess {
 
   implicit private val ec: ExecutionContext = executionContext
-  private val waitForLogs: FiniteDuration = 2.seconds
-  private val logTimeSpanMargin = 1.second
 
   override def run(image: String, args: Seq[String] = Seq.empty[String])(
     implicit transid: TransactionId): Future[ContainerId] = {
@@ -82,35 +71,5 @@ class DockerForMacClient(dockerHost: Option[String] = None)(executionContext: Ex
         case "<no value>" => Future.failed(new NoSuchElementException)
         case stdout       => Future.successful(ContainerAddress("localhost", stdout.toInt))
       }
-  }
-
-  def collectLogs(id: ContainerId, since: Instant, untill: Instant)(implicit transid: TransactionId): Future[String] = {
-    //Add a slight buffer to account for delay writes of logs
-    val end = untill.plusSeconds(logTimeSpanMargin.toSeconds)
-    runCmd(
-      Seq("logs", id.asString, "--since", since.getEpochSecond.toString, "--until", end.getEpochSecond.toString),
-      waitForLogs)
-  }
-}
-
-object DockerForMacLogStoreProvider extends LogStoreProvider {
-  override def instance(actorSystem: ActorSystem): LogStore = {
-    //Logger is currently not passed implicitly to LogStoreProvider. So create one explicitly
-    implicit val logger = new AkkaLogging(akka.event.Logging.getLogger(actorSystem, this))
-    new DockerForMacLogStore(actorSystem)
-  }
-}
-
-class DockerForMacLogStore(system: ActorSystem)(implicit log: Logging) extends DockerToActivationLogStore(system) {
-  private val client = new DockerForMacClient()(system.dispatcher)(log, system)
-  override def collectLogs(transid: TransactionId,
-                           user: Identity,
-                           activation: WhiskActivation,
-                           container: Container,
-                           action: ExecutableWhiskAction): Future[ActivationLogs] = {
-    //TODO Lookup for Log markers to be more precise in log collection
-    client
-      .collectLogs(container.containerId, activation.start, activation.end)(transid)
-      .map(logs => ActivationLogs(logs.linesIterator.toVector))
   }
 }
